PICTOPERFECT-722 - DELETE geeft een 500 We don't do actions when we are busy. period. This resolves a double onclick call with DropDownButton with a trackpad.
git-svn-id: https://ppwcode.googlecode.com/svn/javascript/vernacular/persistence/trunk@7463 463e38ab-4f4e-0410-907a-65d2b3888964

+++ src/persistence/ui/_PersistentObjectEditPane.js
@@ -143,6 +143,10 @@ define(["dojo/_base/declare", "dojo/_base/lang", "ppwcode-vernacular-semantics/u
         this._c_pre(function() {return this.get("crudDao");});
 
         var self = this;
+        if (self.get("presentationMode") !== self.EDIT) {
+          // DropDownMenu gives double click events; in any case, this makes sense to test
+          return null;
+        }
         var po = self.get("target");
         if (!po || !po.get("persistenceId")) {
           return self.close();
@@ -222,6 +226,10 @@ define(["dojo/_base/declare", "dojo/_base/lang", "ppwcode-vernacular-semantics/u
         this._c_pre(function() {return this.get("crudDao");});
 
         var self = this;
+        if (self.get("presentationMode") !== self.EDIT) {
+          // DropDownMenu gives double click events; in any case, this makes sense to test
+          return null;
+        }
         var po = self.get("target");
         var wildExceptions = po && po.get("wildExceptions");
         if (wildExceptions && wildExceptions.isEmpty()) {
@@ -272,6 +280,10 @@ define(["dojo/_base/declare", "dojo/_base/lang", "ppwcode-vernacular-semantics/u
         this._c_pre(function() {return this.get("crudDao");});
 
         var self = this;
+        if (self.get("presentationMode") !== self.EDIT) {
+          // DropDownMenu gives double click events; in any case, this makes sense to test
+          return null;
+        }
         this.set("presentationMode", this.BUSY);
         var po = this.get("target");
         var deletePromise = self.crudDao.remove(po);
